Bug fixes
Box Wallpapers
Stop returning the first box wallpaper regardless of box

EXP:
Changing level/exp wouldn't update stats (EXP needed to be set back to
PKM object)

Drag&Drop:
If the file exists and is locked for editing, the try { } will fail ->
so will delete since the file is locked.
Never seen or heard it happen, but is possible.

+++ MainWindow/Main.cs
@@ -1412,6 +1412,7 @@ namespace PKHeX
                 TB_EXP.Text = PKX.getEXP(Level, Util.getIndex(CB_Species)).ToString();
             }
             changingFields = false;
+            pkm.EXP = Util.ToUInt32(TB_EXP.Text);
             updateStats();
             updateLegality();
         }
@@ -3256,7 +3257,8 @@ namespace PKHeX
             new Thread(() =>
             {
                 Thread.Sleep(500);
-                File.Delete(newfile);
+                if (File.Exists(newfile))
+                    File.Delete(newfile);
             }).Start();
         }
         private void pbBoxSlot_DragDrop(object sender, DragEventArgs e)
